Created error pages
Created an error page to access when typing an invalid address
(or when a page is unavailable) (again)
Created an error page to access
when entering invalid credentials on the login page. (again)
These pages are yet to work as intended.
These pages were removed when reverted certain changes.
There is no need to revert them again.
Reason of reversion will be updated after this commit
(if is ever updated again).